Publish curated HASP release v0.1.34
This release export was generated from the canonical source repo.

Source-Repo: git@github.com:ehmo/hasp.git
Source-Commit: d94097bf4f35d33ac43b4eeabe8f7e6ddff193fd
Source-Tag: v0.1.34

+++ apps/server/internal/runtime/peercred_linux.go
@@ -16,7 +16,7 @@ func realPeerUID(conn net.Conn) (uint32, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint32(cred.Uid), nil
+	return cred.Uid, nil
 }
 
 // realPeerPID returns the PID of the peer connected on conn using
